backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pandas>=2.1.0
openpyxl>=3.1.0
statsmodels>=0.14.0
pmdarima>=2.0.0
matplotlib>=3.8.0
fpdf2>=2.7.0
numpy>=1.24.0
